+++ pages/_app.js
@@ -32,15 +32,6 @@ export default function App({ Component, pageProps }) {
     return Math.trunc(number).toLocaleString("de-DE");
   }
 
-  const [submittedForm, setSubmittedForm] = useState(false);
-  function handleSubmittedForm(value) {
-    setSubmittedForm(value);
-  }
-
-  function handlePrevPage() {
-    setSubmittedForm(false);
-  }
-
   return (
     <>
       <GlobalStyle />
@@ -50,9 +41,6 @@ export default function App({ Component, pageProps }) {
         generalInfo={generalInfo}
         formatNumber={formatNumber}
         menstruationDaysPerYear={menstruationDaysPerYear}
-        handleSubmittedForm={handleSubmittedForm}
-        submittedForm={submittedForm}
-        handlePrevPage={handlePrevPage}
       />
     </>
   );
